# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  notifications.NotificationResponse:
    properties:
      channel:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      read_at:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  notifications.NotificationsListResponse:
    properties:
      limit:
        type: integer
      notifications:
        items:
          $ref: '#/definitions/notifications.NotificationResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
      unread:
        type: integer
    type: object
  notifications.TestEmailRequest:
    properties:
      body:
        type: string
      subject:
        type: string
      to:
        type: string
    required:
    - body
    - subject
    - to
    type: object
  notifications.TestSMSRequest:
    properties:
      body:
        type: string
      to:
        type: string
    required:
    - body
    - to
    type: object
  oauth2.AuthorizeRequest:
    properties:
      client_id:
        type: string
      code_challenge:
        type: string
      code_challenge_method:
        enum:
        - S256
        - plain
        type: string
      redirect_uri:
        type: string
      response_type:
        enum:
        - code
        type: string
      scope:
        type: string
      state:
        type: string
    required:
    - client_id
    - redirect_uri
    - response_type
    type: object
  oauth2.AuthorizeResponse:
    properties:
      code:
        type: string
      state:
        type: string
    type: object
  oauth2.IntrospectRequest:
    properties:
      token:
        type: string
      token_type_hint:
        enum:
        - access_token
        - refresh_token
        type: string
    required:
    - token
    type: object
  oauth2.IntrospectResponse:
    properties:
      active:
        type: boolean
      client_id:
        type: string
      exp:
        type: integer
      iat:
        type: integer
      scope:
        type: string
      token_type:
        type: string
      user_id:
        type: string
    type: object
  oauth2.RevokeRequest:
    properties:
      token:
        type: string
      token_type_hint:
        enum:
        - access_token
        - refresh_token
        type: string
    required:
    - token
    type: object
  oauth2.TokenRequest:
    properties:
      client_id:
        type: string
      client_secret:
        type: string
      code:
        type: string
      code_verifier:
        type: string
      grant_type:
        enum:
        - authorization_code
        - client_credentials
        - refresh_token
        type: string
      redirect_uri:
        type: string
      refresh_token:
        type: string
      scope:
        type: string
    required:
    - client_id
    - grant_type
    type: object
  oauth2.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      scope:
        type: string
      token_type:
        type: string
    type: object
  response.Actor:
    properties:
      client_id:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  response.Meta:
    properties:
      actor:
        $ref: '#/definitions/response.Actor'
      request_id:
        type: string
      timestamp:
        type: string
      version:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/response.ResponseError'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.ResponseError:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@gogin.com
    name: API Support
  description: |-
    A comprehensive REST API built with Go and Gin framework
    This API provides user authentication, profile management, and admin operations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gogin API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Get paginated list of user notifications
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/notifications.NotificationsListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List Notifications
      tags:
      - Notifications
  /notifications/{id}:
    delete:
      description: Delete a notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete Notification
      tags:
      - Notifications
    get:
      description: Get a notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/notifications.NotificationResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Notification
      tags:
      - Notifications
  /notifications/{id}/read:
    put:
      description: Mark a notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Mark Notification as Read
      tags:
      - Notifications
  /notifications/test-email:
    post:
      consumes:
      - application/json
      description: Send a test email via SendGrid
      parameters:
      - description: Email details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notifications.TestEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Test Email
      tags:
      - Notifications
  /notifications/test-sms:
    post:
      consumes:
      - application/json
      description: Send a test SMS via Twilio
      parameters:
      - description: SMS details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/notifications.TestSMSRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                errors:
                  items:
                    $ref: '#/definitions/response.ResponseError'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Test SMS
      tags:
      - Notifications
  /oauth/authorize:
    post:
      consumes:
      - application/json
      description: Request authorization code with PKCE support
      parameters:
      - description: Authorization request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/oauth2.AuthorizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/oauth2.AuthorizeResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: OAuth2 Authorization
      tags:
      - OAuth2
  /oauth/introspect:
    post:
      consumes:
      - application/json
      description: Get information about a token
      parameters:
      - description: Introspect request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/oauth2.IntrospectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/oauth2.IntrospectResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Introspect Token
      tags:
      - OAuth2
  /oauth/revoke:
    post:
      consumes:
      - application/json
      description: Revoke an access or refresh token
      parameters:
      - description: Revoke request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/oauth2.RevokeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Revoke Token
      tags:
      - OAuth2
  /oauth/token:
    post:
      consumes:
      - application/json
      description: Exchange authorization code, refresh token, or client credentials
        for access token
      parameters:
      - description: Token request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/oauth2.TokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/oauth2.TokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: OAuth2 Token
      tags:
      - OAuth2
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...